Organize presentation files in different package
