+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
       android:theme="@style/AppTheme"
       tools:ignore="GoogleAppIndexingWarning"
       >
-    <activity android:name=".SearchableActivity">
+    <activity android:name=".presentation.screen.SearchableActivity">
       <intent-filter>
         <action android:name="android.intent.action.SEARCH" />
       </intent-filter>
@@ -25,7 +25,7 @@
           android:resource="@xml/searchable"
           />
     </activity>
-    <activity android:name=".MainActivity">
+    <activity android:name=".presentation.screen.MainActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
