+++ app/src/main/java/com/hanitacm/weatherapp/presentation/model/DisplayableWeather.kt
@@ -1,4 +1,4 @@
-package com.hanitacm.weatherapp
+package com.hanitacm.weatherapp.presentation.model
 
 data class DisplayableWeather(val description: String,
                               val temperature: Double, val humidity: Int, val location: String)
