+++ app/src/main/java/com/hanitacm/weatherapp/presentation/viewmodel/WeatherViewModel.kt
@@ -1,10 +1,11 @@
-package com.hanitacm.weatherapp
+package com.hanitacm.weatherapp.presentation.viewmodel
 
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
 import android.arch.lifecycle.ViewModelProvider
 import android.util.Log
+import com.hanitacm.weatherapp.presentation.model.DisplayableWeather
 import com.hanitacm.weatherapp.domain.GetWeatherUseCase
 import com.hanitacm.weatherapp.domain.WeatherDomainModel
 import io.reactivex.android.schedulers.AndroidSchedulers
