+++ app/src/main/res/layout/activity_searchable.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchableActivity"
+    tools:context=".presentation.screen.SearchableActivity"
     >
   <android.support.v7.widget.RecyclerView
       android:id="@+id/locations"
